Move ChangeLivestream component into it's own folder
