+++ packages/apolloschurchapp/src/testing-control-panel/changeLivestream.js
@@ -1,5 +1,11 @@
+import React from 'react';
+
+import { Query } from 'react-apollo';
+import { get } from 'lodash';
+
 import { client } from 'apolloschurchapp/src/client';
 import getLiveStream from 'apolloschurchapp/src/live/getLiveStream';
+import TouchableCell from './TouchableCell';
 
 const changeLivestream = ({ isLive }) =>
   client.writeQuery({
@@ -12,4 +18,19 @@ const changeLivestream = ({ isLive }) =>
     },
   });
 
-export default changeLivestream;
+const ChangeLivestream = () => (
+  <Query query={getLiveStream}>
+    {({ data }) => {
+      const isLive = get(data, 'liveStream.isLive', false);
+      return (
+        <TouchableCell
+          handlePress={() => changeLivestream({ isLive: !isLive })}
+          iconName={isLive ? 'pause' : 'play'}
+          cellText={`${isLive ? 'End' : 'Start'} The Livestream`}
+        />
+      );
+    }}
+  </Query>
+);
+
+export default ChangeLivestream;
